LayoutTests/imported/w3c/web-platform-tests/css/css-align/abspos/align-self-vlr-ltr-htb-expected.txt




FAIL .item 1 assert_equals:
<div class="container">
  <div class="item" style="align-self: start;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
FAIL .item 2 assert_equals:
<div class="container">
  <div class="item" style="align-self: center;" data-expected-width="20" data-offset-x="10"></div>
</div>
width expected 20 but got 40
FAIL .item 3 assert_equals:
<div class="container">
  <div class="item" style="align-self: end;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
FAIL .item 4 assert_equals:
<div class="container">
  <div class="item" style="align-self: baseline;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
FAIL .item 5 assert_equals:
<div class="container">
  <div class="item" style="align-self: last baseline;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
FAIL .item 6 assert_equals:
<div class="container">
  <div class="item" style="align-self: self-start;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
FAIL .item 7 assert_equals:
<div class="container">
  <div class="item" style="align-self: self-end;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
PASS .item 8
FAIL .item 9 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: start;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
FAIL .item 10 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: center;" data-expected-width="20" data-offset-x="10"></div>
</div>
width expected 20 but got 40
FAIL .item 11 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: end;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
FAIL .item 12 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: baseline;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
FAIL .item 13 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: last baseline;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
FAIL .item 14 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: self-start;" data-expected-width="20" data-offset-x="20"></div>
</div>
width expected 20 but got 40
FAIL .item 15 assert_equals:
<div class="container">
  <div class="item rtl" style="align-self: self-end;" data-expected-width="20" data-offset-x="0"></div>
</div>
width expected 20 but got 40
PASS .item 16
